+++ src/main/scala/sc4pac/api/message.scala
@@ -307,7 +307,12 @@ case class ServerStatus(sc4pacVersion: String, osVersion: String, javaVersion: S
 
 case class ProfileName(name: String) derives UP.ReadWriter
 case class ProfileIdObj(id: ProfileId) derives UP.ReadWriter
-case class ProfilesList(profiles: Seq[JD.ProfileData], currentProfileId: Option[ProfileId], profilesDir: java.nio.file.Path) derives UP.ReadWriter
+case class ProfilesList(profiles: Seq[ProfilesList.ProfileData2], currentProfileId: Option[ProfileId], profilesDir: java.nio.file.Path) derives UP.ReadWriter
+object ProfilesList {
+  case class ProfileData2(id: ProfileId, name: String, pluginsRoot: Option[java.nio.file.Path] = None) derives UP.ReadWriter {
+    def toProfileData: JD.ProfileData = JD.ProfileData(id = id, name = name)
+  }
+}
 
 case class VariantsList(variants: Map[String, VariantsList.Item])
 object VariantsList {